+++ src/main/java/org/example/Entity/Rectangle.java
@@ -2,10 +2,10 @@ package org.example.Entity;
 
 public class Rectangle extends Figure {
 
-    private final double width;
-    private final double height;
+    private final int width;
+    private final int height;
 
-    public Rectangle(String color, double width, double height) {
+    public Rectangle(String color, int width, int height) {
         super(color);
         this.width = width;
         this.height = height;
